# Conflicts:
#	src/gamedata/fonts/font.cpp
#	src/gamedata/fonts/fontinternals.h
#	src/gamedata/fonts/specialfont.cpp
#	src/gamedata/fonts/v_font.h
#	src/gamedata/textures/texturemanager.cpp
#	src/gamedata/textures/textures.h
#	src/maploader/renderinfo.cpp
#	wadsrc/static/language.enu
#	wadsrc/static/zscript/menu/optionmenu.txt

+++ src/gamedata/fonts/fontinternals.h
@@ -37,6 +37,6 @@ extern TArray<PalEntry> TranslationColors;
 extern uint16_t lowerforupper[65536];
 extern uint16_t upperforlower[65536];
 
-void RecordTextureColors (FTexture *pic, uint8_t *colorsused);
+void RecordTextureColors (FTexture *pic, uint32_t *colorsused);
 bool myislower(int code);
 int stripaccent(int code);
